pyproject.toml
[project]
name = "first-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai-agents>=0.2.2",
    "prompt-toolkit>=3.0.51",
    "pygments>=2.19.2",
    "python-dotenv>=1.1.1",
    "rich>=14.0.0",
    "typer>=0.16.0",
]
